+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DropByPredicateInclusiveSuite.scala
@@ -38,7 +38,7 @@ object DropByPredicateInclusiveSuite extends BaseOperatorSuite {
   def observableInError(sourceCount: Int, ex: Throwable) = {
     require(sourceCount > 0, "sourceCount should be strictly positive")
     Some {
-      val o = createObservableEndingInError(Observable.range(1, sourceCount + 2), ex)
+      val o = createObservableEndingInError(Observable.range(1, sourceCount.toLong + 2), ex)
         .dropWhileInclusive(_ < 1)
 
       Sample(o, count(sourceCount), sum(sourceCount), 0.seconds, 0.seconds)
