+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DropByPredicateWithIndexSuite.scala
@@ -42,7 +42,7 @@ object DropByPredicateWithIndexSuite extends BaseOperatorSuite {
   def observableInError(sourceCount: Int, ex: Throwable) = {
     require(sourceCount > 0, "sourceCount should be strictly positive")
     Some {
-      val o = createObservableEndingInError(Observable.range(1, sourceCount + 2), ex)
+      val o = createObservableEndingInError(Observable.range(1, sourceCount.toLong + 2), ex)
         .dropWhileWithIndex((e, idx) => idx == 0)
 
       Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
