+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/MapParallelUnorderedSuite.scala
@@ -297,7 +297,7 @@ object MapParallelUnorderedSuite extends BaseOperatorSuite {
     val totalCount = Platform.recommendedBatchSize * 4
 
     Observable
-      .range(0, totalCount)
+      .range(0, totalCount.toLong)
       .doOnNext(_ => Task { initiated += 1 })
       .mapParallelUnordered(parallelism = 4)(x => Task.evalAsync(x))
       .unsafeSubscribeFn(new Observer[Long] {
@@ -332,7 +332,7 @@ object MapParallelUnorderedSuite extends BaseOperatorSuite {
     val totalCount = Platform.recommendedBatchSize * 4
 
     Observable
-      .range(0, totalCount)
+      .range(0, totalCount.toLong)
       .doOnNext(_ => Task { initiated += 1 })
       .mapParallelUnordered(parallelism = 4)(x => Task.evalAsync(x))(
         OverflowStrategy.DropNew(Platform.recommendedBatchSize))
