+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/TakeByPredicateSuite.scala
@@ -49,7 +49,7 @@ object TakeByPredicateSuite extends BaseOperatorSuite {
     require(sourceCount > 0, "sourceCount should be strictly positive")
     Some {
       val ex = DummyException("dummy")
-      val o = createObservableEndingInError(Observable.range(1, sourceCount + 1), ex)
+      val o = createObservableEndingInError(Observable.range(1, sourceCount.toLong + 1), ex)
         .takeWhile(_ <= sourceCount * 2)
 
       Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
