+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/Zip2Suite.scala
@@ -29,7 +29,7 @@ import scala.concurrent.duration._
 object Zip2Suite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Some {
     val o1 = Observable.range(0L, sourceCount.toLong)
-    val o2 = Observable.range(0, sourceCount + 2)
+    val o2 = Observable.range(0, sourceCount.toLong + 2)
 
     val o = Observable.zipMap2(o1, o2) { (x1, x2) =>
       x1 + x2
@@ -52,7 +52,7 @@ object Zip2Suite extends BaseOperatorSuite {
 
   def brokenUserCodeObservable(sourceCount: Int, ex: Throwable) = Some {
     val o1 = Observable.range(0L, sourceCount.toLong)
-    val o2 = Observable.range(0, sourceCount + 100)
+    val o2 = Observable.range(0, sourceCount.toLong + 100)
 
     val o = Observable.zipMap2(o1, o2) { (x1, x2) =>
       if (x2 < sourceCount - 1) x1 + x2 else throw ex
