+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/Zip4Suite.scala
@@ -25,8 +25,8 @@ object Zip4Suite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Some {
     val o1 = Observable.range(0L, sourceCount.toLong).executeAsync
     val o2 = Observable.range(0, sourceCount.toLong + 1).executeAsync
-    val o3 = Observable.range(0, sourceCount + 2).executeAsync
-    val o4 = Observable.range(0, sourceCount + 3).executeAsync
+    val o3 = Observable.range(0, sourceCount.toLong + 2).executeAsync
+    val o4 = Observable.range(0, sourceCount.toLong + 3).executeAsync
 
     val o = Observable.zipMap4(o1, o2, o3, o4)(_ + _ + _ + _)
     Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
@@ -47,7 +47,7 @@ object Zip4Suite extends BaseOperatorSuite {
 
   def brokenUserCodeObservable(sourceCount: Int, ex: Throwable) = Some {
     val o1 = Observable.range(0L, sourceCount.toLong).executeAsync
-    val o2 = Observable.range(0, sourceCount + 100).executeAsync
+    val o2 = Observable.range(0, sourceCount.toLong + 100).executeAsync
     val o3 = Observable.range(0L, sourceCount.toLong).executeAsync
     val o4 = Observable.range(0L, sourceCount.toLong).executeAsync
