+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/Zip6Suite.scala
@@ -25,10 +25,10 @@ object Zip6Suite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Some {
     val o1 = Observable.range(0L, sourceCount.toLong).executeAsync
     val o2 = Observable.range(0, sourceCount.toLong + 1).executeAsync
-    val o3 = Observable.range(0, sourceCount + 2).executeAsync
-    val o4 = Observable.range(0, sourceCount + 3).executeAsync
-    val o5 = Observable.range(0, sourceCount + 4).executeAsync
-    val o6 = Observable.range(0, sourceCount + 5).executeAsync
+    val o3 = Observable.range(0, sourceCount.toLong + 2).executeAsync
+    val o4 = Observable.range(0, sourceCount.toLong + 3).executeAsync
+    val o5 = Observable.range(0, sourceCount.toLong + 4).executeAsync
+    val o6 = Observable.range(0, sourceCount.toLong + 5).executeAsync
 
     val o = Observable.zipMap6(o1, o2, o3, o4, o5, o6)(_ + _ + _ + _ + _ + _)
     Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
@@ -51,7 +51,7 @@ object Zip6Suite extends BaseOperatorSuite {
 
   def brokenUserCodeObservable(sourceCount: Int, ex: Throwable) = Some {
     val o1 = Observable.range(0L, sourceCount.toLong).executeAsync
-    val o2 = Observable.range(0, sourceCount + 100).executeAsync
+    val o2 = Observable.range(0, sourceCount.toLong + 100).executeAsync
     val o3 = Observable.range(0L, sourceCount.toLong).executeAsync
     val o4 = Observable.range(0L, sourceCount.toLong).executeAsync
     val o5 = Observable.range(0L, sourceCount.toLong).executeAsync
